changed PasswordEncoder and divided userService
