+++ hw11-spring-security-auth/src/main/java/ru/otus/hw/services/UserService.java
@@ -1,9 +1,8 @@
 package ru.otus.hw.services;
 
-import org.springframework.security.core.userdetails.UserDetailsService;
 import ru.otus.hw.dto.UserDto;
 
-public interface UserService extends UserDetailsService {
+public interface UserService {
 
     UserDto findByUsername(String username);
 }
