+++ hw11-spring-security-auth/src/main/resources/db/changelog/data/2024-10-20--0002-users.sql
@@ -3,4 +3,5 @@
 --changeset IlyaLeshin:2024-10-20-003-users
 merge into users(username, password)
 key(username)
-values ('user1', 'password1'), ('user2', 'password2');
+values ('user1', '$2a$12$KQZtdFV4vtml5IKzvYpWreewsKvl6cWjVylmJ025Aqz/T8VS8UDSi'),
+('user2', '$2a$12$2s3mj6h7PFcws1TydeJRDeapRMe/pgxAfqh0HIJ0c13gUCgK0PC2K');
